use var path instead of /tmp, also directory remover handle no removal needed

+++ src/Part1/Chapter1/iterator.php
@@ -8,10 +8,16 @@ use Book\Part1\Chapter1\IteratorFun\Config;
 use Book\Part1\Chapter1\IteratorFun\DirectoryRemover;
 use Book\Part1\Chapter1\IteratorFun\FileCreator;
 use Book\Part1\Chapter1\IteratorFun\FilterBlueFiles;
+use RuntimeException;
 
 require __DIR__ . '/../../../vendor/autoload.php';
 
-$config = new Config('/tmp/iterator-fun', 'foo/bar/baz', 'doo/dar/daz');
+const TMP_DIR = __DIR__ . '/../../../var/';
+if (!is_dir(TMP_DIR) && mkdir(TMP_DIR, 0777, true) && !is_dir(TMP_DIR)) {
+    throw new RuntimeException(sprintf('Directory "%s" was not created', TMP_DIR));
+}
+
+$config = new Config(TMP_DIR . '/iterator-fun', 'foo/bar/baz', 'doo/dar/daz');
 
 /* First we use the DirectoryRemover to clean up any previous run */
 (new DirectoryRemover())->removeDir($config->getBaseDir());
